Reduce reliance on posix_force_include.h
Move functionality to respective header files.

b/302740372

Test-On-Device: true
