+++ starboard/shared/win32/posix_emu/include/strings.h
@@ -12,16 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <string.h>
+#ifndef STARBOARD_SHARED_WIN32_POSIX_EMU_INCLUDE_STRINGS_H_
+#define STARBOARD_SHARED_WIN32_POSIX_EMU_INCLUDE_STRINGS_H_
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-char* strdup(const char* s1) {
-  return _strdup(s1);
-}
+// https://pubs.opengroup.org/onlinepubs/9699919799/functions/strcasecmp.html
+int strcasecmp(const char* s1, const char* s2);
+
+// https://pubs.opengroup.org/onlinepubs/9699919799/functions/strncasecmp.html
+int strncasecmp(const char* s1, const char* s2, size_t n);
 
 #ifdef __cplusplus
 }
-#endif
+#endif  // __cplusplus
+
+#endif  // STARBOARD_SHARED_WIN32_POSIX_EMU_INCLUDE_STRINGS_H_
